+++ s2jdbc-it/src/test/java/org/seasar/extension/jdbc/it/SqlSelectPagingTest.java
@@ -55,8 +55,8 @@ public class SqlSelectPagingTest extends S2TestCase {
         List<Employee> list = jdbcManager.selectBySql(Employee.class, sql)
                 .offset(3).getResultList();
         assertEquals(11, list.size());
-        assertEquals(4, list.get(0).employeeId.intValue());
-        assertEquals(14, list.get(10).employeeId.intValue());
+        assertEquals(4, list.get(0).employeeId);
+        assertEquals(14, list.get(10).employeeId);
     }
 
     /**
@@ -67,8 +67,8 @@ public class SqlSelectPagingTest extends S2TestCase {
         List<Employee> list = jdbcManager.selectBySql(Employee.class, sql)
                 .limit(3).getResultList();
         assertEquals(3, list.size());
-        assertEquals(1, list.get(0).employeeId.intValue());
-        assertEquals(3, list.get(2).employeeId.intValue());
+        assertEquals(1, list.get(0).employeeId);
+        assertEquals(3, list.get(2).employeeId);
     }
 
     /**
@@ -89,8 +89,8 @@ public class SqlSelectPagingTest extends S2TestCase {
         List<Employee> list = jdbcManager.selectBySql(Employee.class, sql)
                 .offset(3).limit(0).getResultList();
         assertEquals(11, list.size());
-        assertEquals(4, list.get(0).employeeId.intValue());
-        assertEquals(14, list.get(10).employeeId.intValue());
+        assertEquals(4, list.get(0).employeeId);
+        assertEquals(14, list.get(10).employeeId);
     }
 
     /**
@@ -101,8 +101,8 @@ public class SqlSelectPagingTest extends S2TestCase {
         List<Employee> list = jdbcManager.selectBySql(Employee.class, sql)
                 .offset(0).limit(3).getResultList();
         assertEquals(3, list.size());
-        assertEquals(1, list.get(0).employeeId.intValue());
-        assertEquals(3, list.get(2).employeeId.intValue());
+        assertEquals(1, list.get(0).employeeId);
+        assertEquals(3, list.get(2).employeeId);
     }
 
     /**
@@ -113,7 +113,7 @@ public class SqlSelectPagingTest extends S2TestCase {
         List<Employee> list = jdbcManager.selectBySql(Employee.class, sql)
                 .offset(3).limit(5).getResultList();
         assertEquals(5, list.size());
-        assertEquals(4, list.get(0).employeeId.intValue());
-        assertEquals(8, list.get(4).employeeId.intValue());
+        assertEquals(4, list.get(0).employeeId);
+        assertEquals(8, list.get(4).employeeId);
     }
 }
